+++ src/checkout/component.js
@@ -10,7 +10,7 @@ import type { CrossDomainWindowType } from 'cross-domain-utils/src';
 import { isDevice, request, getQueryParam, redirect as redir, patchMethod,
     setLogLevel, getSessionID, getBrowserLocale, supportsPopups, memoize,
     getDomainSetting, getScriptVersion, getButtonSessionID, isPayPalDomain,
-    isEligible, getCurrentScript } from '../lib';
+    isEligible, getCurrentScriptUrl } from '../lib';
 import { config } from '../config';
 import { ENV, FPTI, PAYMENT_TYPE, CHECKOUT_OVERLAY_COLOR, ATTRIBUTE } from '../constants';
 import { onLegacyPaymentAuthorize } from '../compat';
@@ -539,15 +539,10 @@ export let Checkout : Component<CheckoutPropsType> = create({
             sendToChild: false,
             def:         () => {
                 return btoa(JSON.stringify({
-                    url: getCurrentScript()
+                    url: getCurrentScriptUrl()
                 }));
             }
         }
-    },
-
-    autoResize: {
-        width:  false,
-        height: false
     },
 
     get dimensions() : { width : string, height : string } {