+++ webpack.config.js
@@ -165,7 +165,16 @@ export function getWebpackConfig({
         resolve: {
             extensions: [ '.js', '.jsx' ]
         },
-        plugins
+        plugins,
+        node: {
+            console:      false,
+            global:       false,
+            process:      false,
+            __filename:   false,
+            __dirname:    false,
+            Buffer:       false,
+            setImmediate: false
+        }
     };
 
     if (src) {